Non-Compiling Version of DMA Module
This uses RPIO v2 on the RPi 3B+ to do DMA-timed PWM.

As written, it uses a 2ms subcycle, with pulse increments of 1us. This gives us 2000 steps of dimming, at 500Hz. Could be tweaked a bit, but will never quite be as good as a dedicated PWM controller like the adafruit. We’ve only got microsecond resolution, and have to trade frequency for dimming.

As it is, we have just shy of 11 bits of dimming resolution, at 500Hz. To get to 12 bits, we need to run at around 250Hz, which starts getting rather low.

I do have an idea for a trick that could get me 500Hz at low intensities, and 250Hz at high intensities, while giving me more resolution though. Will need to experiment with it.