Enable DMA PWM Module
This makes it possible to use the DMA PWM module.

When on a platform that doesn’t have RPIO, then it will fail if you call pwm.Startup(MODE_DMA) when it attempts to load the package.

The Simulator mode is always available, and this change makes it possible to use the simulator mode on a non-RPi machine, but then deploy to an RPi to check the changes against live hardware.