Write Initial Controller Code
This is an early version of the controller code. It doesn’t handle priorities or anything else yet, just sequential RampBehavior objects.

The idea here is:

- LightController is in charge of scheduling.
- Events will be added to the LightController which trigger at certain times of day, and perform actions. This can include configuring and adding Behaviors to the LightController.
- Behaviors when added to the LightController are run in a form of priority order. They have start and stop datetimes.

The common behavior is the RampBehavior. This changes the light settings from a start state to an end state over the period of time assigned to it. On paper, it should be able to handle being started partway through its range.

Did some initial testing with test.py that reuses a RampBehavior and runs it twice. Not quite as nice as the up/down ramp before, but does a decent test of the RampBehavior and LightController for now.