Add Schedule Parser
Now it is time to start getting serious, add the Schedule parser to the codebase, and start writing the preview.py script.

This new test script is meant to read a hardware config, a schedule config, and then run the “Preview” mode on it. This is a custom behavior that will be added to the controller that will run through a schedule given to it, using RampBehaviors and waits.

It isn’t implemented, so for now, preview.py just reports back what the schedule looks like once parsed. So far it looks pretty good.