Implement Gamma Correction
Linear ramps of LEDs don’t look quite right to the human eye. We set something to “50%” and it looks a bit brighter than that, because we don’t track light linearly.

Add in a basic gamma correction when SetBrightness is called on a channel. Internally, pass on this corrected luminance value to the actual PWM driver code.