Implement Preview Script
This implements a basic preview script. It compresses a schedule from 24 hours down to 1 minute and plays it out on the channels very quickly.

It has a couple gaps, and doesn’t fully validate/sort the events in the schedule, which can cause it to hang because it assumes everything is already sorted, but that can be fixed later.