Add Scheduler Daemon
Here we go, a daemon that runs the full scheduler. For now it uses the test configuration like the other scripts.

This change is a little big, and includes the following:
- Rename RampBehavior to LightBehavior
- Refactor the Channel information used in various places to be a dictionary that uses the channel token as a key. This is important to make it a bit easier to map a set of deltas to apply to a channel, to the actual channels to have them applied to. i.e. We no longer depend on ordering. Takes more CPU time, but is a good tradeoff since it makes things more reliable.
- Adds in support for cron events. These are created based on the schedule, and fire based on the cron timings. They apply new behaviors to the controller when they fire based on what was calculated to be correct for their time slot.
- Move calculating a range of brightness values to the Schedule, so the scheduler and preview tools can both use it.

Needs testing against a real fish tank though. To see how well it works. Also starting to see how much code duplication exists right now.