Implement Priority Queue
This implements a priority queue for the light controller, and replaces the multiple behavior jobs with a single job. This job will pause/resume as behaviors are added and removed, as appropriate.

Also added constants for the priority queue for the general light ramp and the preview. Also threw in an override priority, just so it exists as a concept for now, even though it is unused.

The idea of priorities makes more sense in the context of the schedule daemon, as it allows events like thunderstorms or lunar cycles to work as overrides to the expected behavior with a rather simple approach. It is getting implemented now, because there’s a bug in the scheduler where a behavior added right when the previous one ends winds up getting blocked and never runs.