Tweaks to Priority Queue Behavior
There’s a couple issues still when a behavior goes onto a different behavior that this addresses:
- Ensure that the final tick of a behavior is allowed to fire.
- Make sure that if we wind up calling into the handler before it is time for the new behavior to happen, then don’t actually trigger the behavior.
- Fixed a bug where we would bind the brightness to a range, but then use the unbound value anyways, which is generally a bad thing to do.

To help a bit catch issues like this in the future, the test script now runs a couple of quick ramps. One with a gap, and one with a 1 second overlap. It may need to also run a ramp that has no overlap at all.