Implement Adafruit PCA9685 Driver
Now that this is written, it is now possible to support 16 channels, with 12-bit resolution, and up to 1440Hz PWM frequency. For those who need more that then 2 built in channels, this is a pretty good solution.

This change also enables the driver, and adds an example configuration for an 8 channel setup. 

It would be nice to be able to stagger the pulses in the future, but that can be left for a later change.