Further Fixes for Adafruit Driver
It turns out that I was writing 4096 to the LED_OFF register for the channel. This has a special meaning which is: Always off.

So we'd hit 100% light, and turn off instead. This is a sort of off by one error because the driver set the range to 0-4096 instead of 0-4095.