Cleanup Adafruit Driver
This changes the Adafruit driver to use the same numbering scheme as on the board. So instead of 1-16, it uses 0-15. This will make it easier for folks to wire things up and configure it.

Also added an example schedule for the Adafruit. Not sure why that wasn’t checked in sooner.