Implement Extra Special Scheduler
APScheduler hasn’t been working out. There’s race conditions because tasks aren’t guaranteed to be in any specific order. It takes 4 seconds to load on the Raspberry Pi Zero, and it sucks down CPU like a fire hose.

This scheduler is a simpler event loop thread. When it is woken up, or when it wakes up because of an event it already knows about, it goes through a loop handling events and behaviors in that order. Because I know the order, things get a bit easier. The race conditions are gone because of the event handler loop being more consistent.

Instead of 28% CPU at 24 updates/second, it uses under 14% CPU. Much better and close to the range I was aiming for.

No more boot lag either.

Also update .gitignore since macOS keeps adding DS_Store files.