Fix Controller Start Before First Event of the Day
If we are before the first event of the day when the service starts, it won't configure the state properly. This is because it sets the "next datetime" to the last event "today" instead of "yesterday". Account for this when setting the next event datetime.

While we are at it, clean up some of the ordering and logging to make things a bit clearer.