Further Fixes to "Start Before First Event" Issue
There's another situation where things don't work right when starting the daemon before the first event of the day.

With the first attempt at the fix, when it fired the last event of the previous day (to set the current light level), it calculated the next event datetime. But instead of using the date of the event firing, it uses the current date. Which meant it would wait until the next day.