Experiment With RPiLight on Swift
This has the potential to greatly shrink the amount of CPU time needed to run the RPiLight project compared to the python version which can use up to 14% of the CPU on the Pi Zero (@24 fps). Some early tests suggest this should be around 2%, and I can jettison the pigpiod dependency, which costs another 5-7% CPU.

This is an early build that creates a simulated channel and writes output as a sort of validation test. Early testing shows there’s a bug in SwiftyGPIO I need to fix, and will need to leverage a fork to do the experimentation there.

The rest should go pretty quickly, minus the PCA9685 driver, really.