Load Configurations in Swift
Write a loader of configurations, based on the python version, but written in Swift. It uses the JSONSerializer in Foundation to read the file into memory. But from there it does its own deserialization. Unfortunately, instead of throwing useful errors, it returns nil if something is wrong. So it still needs to provide a good error message. Something the python version also lacks.

This also needs a way to validate the state of the configuration to ensure it is valid. Many places, there is some amount of confirmation on reading, but do we want all validation to happen on load, and if it fails to load throw? Or do we want to have a validation pass afterwards?

The reason this last bit is important is because certain parts of the configuration depend on others. I have to have the channel list from the PWM module to confirm that tokens used are valid, for example. I need to know what channels are used to know if a schedule’s events have valid tokens.

So I think the guidance will probably be:
- If it is bounded to a fixed set of things, do it on load.
- A validation should run afterwards that checks that the configuration is consistent with itself.