Implement LightController
This is the Swift version of the LightController. It handles schedules almost as well as the python version. But it does have a couple incomplete gaps that need to be addressed:

- It doesn’t calculate any real interval, it just runs 100 times a second for every behavior.
- There is some cleanup around what type we want to hang calcNextDate off of. Do we make it the Timed protocol, or the DateComponents type itself?

There’s also some interesting stuff happening around dispatch sources. When changing the schedule of the timer source, it seems to shut itself off if scheduling it while it is running. Doesn’t do this for the one-shot scheduling. That’s a rather annoying behavior there. These have both been deprecated for Swift 4, so who knows. Perhaps it makes it a little easier.