Unify calcNextDate Implementations
Because Linux doesn’t have an implementation of enumerateDates() that Calendar.nextDate() depends on, we need to implement a work around for Linux platforms. It won’t be as accurate, which is a shame, but we can give it a shot.

To make this easier, this change unifies the calcNextDate implementations into a single copy attached to DateComponents so it is possible to fix this once for Linux going forward.