Initial Attempt at Adjustable Behavior Intervals
This doesn’t do a whole lot at the moment, other than do the calculation, but we can start look at how frequently we recalculate things and how it affects performance. The hope here is that we get a bit of a ramp on update frequency while we produce a ramp of brightness.