Add Early Command Lines
Turns on the ability to pass in a verbosity flag, and to specify the config file being used. This should help pave the way for re-adding support for previews and test ramps. It’s possible that we are now at the point where we really only need previews, and can use that to generate specific test ramps.