Improve Time Formatting in Logging
A lot of the date logging used the date directly without formatting it first. Instead, lets do some formatting.

Going forward, we really should be accounting for how many of these formatters we wind up creating. It could be a bit messy having these copies all over the place.

Perhaps we need a sort of support for formatting dates in the Logger instead that can be queried?