Implement Previewer
Implements the core functionality of a previewer. It shrinks 24 hours down to a single minute, and plays back the provided schedule during that time.

There is a bug here where it starts from the first event of the day, and loops around to that point in time again. It doesn’t start from midnight, which is not ideal but will work for the short term.

This also fixes/changes quite a few smaller things discovered:
- Fixes an issue when calculating the brightness that meant downward ramps jumped to the end state. This was bad bounding of the brightness to the LightRange.
- Fixes around behavior scheduling to make it more consistent, and similar to what was being done by the python version.
- Stop overloading the logging functions with the closure versions. Those are now “with<Level>” (e.g. withDebug).
- calcNextDate now returns an optional. A couple fewer ‘!’ in the code.