Further Logging Improvements
Switches to StdErr for output to ensure that it reaches the journal. It looks a bit like there’s an issue with StdOut not making it to the journal, despite being configured for it. But that needs further research.

Also adds some tweaks to what is currently logged around new behaviors being added, and the config URL on errors.