Fix Work Queue Ordering
When a behavior is to be removed, if there is a block in the dispatch queue already which will modify the behavior, RPiLight will wind up removing the new behavior, even though the intent was to remove the old one (which is already removed).

It should check to see if it is the correct behavior before removing it.