Fix Issue of Setting Brightness on Start
This issue was getting masked because I was working with the PCA9685, which keeps the previous PWM running so long as power isn’t cut. Even on a reboot, it will keep things running.

But we should now properly be handling being started in the middle of the schedule again.