Support Minimum Intensity
This is an optional configuration tool that lets someone configure the off setting for a particular light channel. This is mostly useful for all-in-one lights being controlled that don’t turn off at 0% duty cycle, but some other value close to it.

An example is the Twinstar 600E which starts turning off at about 0.2%