Fixup the Minimum Intensity Behavior
The first implementation wasn’t actually correct. The problem with it was that it would shift the meaning of brightness and intensity in ways that made no sense.

This version acts like a real cut off. When creating ramps, we treat the cut-off as a sort of ‘zero’ point for the channel so that an on/off ramp will turn the lights on or off at the correct time, but leaves the meanings of any absolute intensity alone. Unlike the first attempt.

TODO: Still need to report back the correct setting value when we clip, while avoiding recursion in didSet.