Support Multiple Raspberry Pi Boards
The bad news is that auto-detection of the board didn’t play out as well as I would have liked. It’s probably doable, but it’s also a bit more important to unblock targeting multiple boards sooner.

So for now, it’s required that you specify which board you have in the configuration, which it will use to configure the hardware I/O properly.

Mostly only matters for the hardware GPIO.