Implement Best Guess for Raspberry Pi Board
Turns out that pigpio just uses the CPU type to figure out what it is running on. It’s not ideal, but it does mean folks aren’t required to set their board type unless we fail to try to detect it, or they need a specific override down the road that isn’t known about today. SwiftyGPIO only really needs to know about the ARMv6 vs ARMv7 difference for GPIO addresses. Otherwise they are the same. I2C is the same across all Raspberry Pi models. So this really is “good enough” for our purposes.

Also, since we can autodetect desktop easily enough this way (look for x86 or x86_64), go ahead and do that as well. We can throw more appropriate errors when you attempt to use anything other than the simulated module on desktop.