Add libpthread-workqueue-dev to Dependencies
Found this while setting up a clean slate Pi Zero install. Not sure how this got missed, but this dependency wasn’t getting installed, but we’d fail on a clean Raspbian Lite image because of this.

Once adding this to the dependencies, the whole process on a clean Raspbian Lite looks like:

- sudo apt-get install git
- git clone <RPiLight>
- ./install.sh full

Yup, three steps once you can log into Raspbian to get an installed build. And we are now at the point where everything not handled by apt now lives in /opt, making cleanup a lot easier as well (/opt/rpilight and /opt/swift).

Possibly the next step around cleanup of the install is to have a more sane layout for /opt/rpilight.