Clean Up Dependencies
Switch to using the official Moderator repository, now that they've taken the pull request to support Swift 3 & 4 simultaneously. 

Swift 4 also has a really nice ".package(url:from:)" notation to specify dependencies. I'd like to be able to use that, so let's abuse the fact that the Package file is a swift file, and backport it as an extension.