Refactor of Configuration (DNR)
This refactors the configuration model so that it represents the new format we want to use going forward for the JSON configuration file that greatly improves readability of it.

It also adds a bevy of testing around the configuration to confirm the behavior is what we want it to be. Should help cut down on new bugs introduced by this change.

Bad news is that it breaks the daemon for the time being. The rest of the refactor of the LightController and the Previewer needs to be done to start using this new configuration.