Major Refactor Work (LightController)
This completes the bulk of the refactoring work. It refactors the LightController to use the new configuration information, and does some other complicated things to how we

- Light Controller: Behavior
The light controller contains a behavior which helps configure/run the controller. It is a way to inject behaviors into the controller. The default behavior follows the Layers’ light levels using the time clock that the controller is using, which is the system clock. The Preview is now implemented as a behavior that runs from midnight to midnight by translating the controller’s clock into an accelerated one that runs the whole 24-hour range of the Layer in 1 minute.

These behaviors enable previews in a more compact way, but they can also enable thunderstorms. Thunderstorms will be a behavior that temporarily takes over, and because it takes over the update/refresh logic, it can do the sort of specific timing required to make lighting strikes look good.

- Light Controller: Channel Controllers
These controllers work on a single channel. They use the controllers layers to update the channel. This means each channel can now be on their own separate schedule.

Layers will help enable lunar/thunderstorm behavior by allowing multiple layers to multiply blend together. The base layer is the user’s light schedule, but a lunar cycle layer can do something like cut the user’s light schedule by 50% during a certain period of time, using the multiply blend behavior.

Previews have been tested, and they fix an issue that’s been ongoing with previews. General behavior seems to be correct right now as well.

Tests have been added, and have roughly 83% coverage on the new Core module.

Some outstanding issues:
- Lost the ability to invalidate and change the timer that causes the updates. That needs to be fixed before this can be merged back into the master branch.
- One-Shot updates haven’t been tested/fixed.

We’ve lost the concept of events that fire during the schedule. This isn’t a complete loss, but they will likely come back in a more limited way as a way to dynamically trigger updates for Lunar Cycles and Thunderstorms, as they both create/leverage dynamic layers that need to be updated periodically.