Fix up Dynamic Interval Behavior
There were some issues which caused RPiLight to go to sleep when it really shouldn’t have.

Let’s do some work to go back to behavior similar to before the refactor. It does require a little touchup to how we get information about the start/end states of a segment, but this is actually a little cleaner than before. So no harm no foul.