Add EventController / LunarCycleController
In order to handle the dynamic layer required (a new layer each day), there needs to be something that fires at some regular time to calculate that dynamic layer and sets it on all the channels.

This is an initial implementation, which also needs unit tests badly. This should come in a follow-up commit.