Improve Packaging Output
This work is to make it easier to host packages in a custom Gemfury repo. armv6 vs armv7 is included in the package name to be repo friendly. Version strings should now be more complete when building a package newer than the tag it is based on. This will let me package pre-release builds more easily for myself or others.