Fix Bug w/Schedule Segment Merging
Previously, layers were being merged the same way that channels were being merged. They both were technically wrong.

Channels want to carve out the slice of time that represents when the last layer change happened, to the next layer change, relative to the date provided. Behaviors want the same thing. They differ on what the brightness range should be.

When merging layers, we want this to represent the actual light levels. So we should interpolate new start/end brightness levels and then multiply them together.

When merging channels, we want the range that represents the largest delta between the two. We still interpolate, but then we just pick the biggest total difference between the two.

This should fix some scheduling issues that the refactor introduced.