Fix Interval for testForceStop
The fix here is to up the refresh interval high enough that we should timeout on the expectations first.

The original mock behavior was that it would stop itself after 200 ms, but depending on how the queue is setup, the force stop may just land shortly after that point. This should prevent the race condition from showing up during the test, which just wants to make sure that the controller stopped by the .stop() call, and not the updates.