Another Form of Fix for Intervals
This fix is probably more complete, as it changes the formula for the interval:

We calculate the number of updates we want during the duration of the segment, rounding it up to the next whole number. Then divide the duration by that number to get the desired interval. Then bind the interval to the range bounded by 10 milliseconds, and the duration of the segment.

This means that any interval we calculate should be an even divisor of the duration, by forcing the update count to be a whole number. There will never be an interval that should cause it to overflow, assuming everything uses floating point. However, because the interval is in milliseconds, there are fractional values that get truncated.

There should be a way to calculate how much error is possible here, but I’m not able to work it out at the moment.