Use DispatchTimeInterval in LightBehaviorUpdate
There’s a bit of a problem when calculating repeating intervals after the refactor. Before, each segment had a fixed event associated with it. That would prevent weird intervals from causing lag when switching segments.

After the refactor, one extra update needed to happen to switch to a new segment. This update could be very delayed as there was nothing preventing something like update intervals larger than 1/2 the entire segment’s length. A fix was to divide the segment’s length by the number of desired updates. This reduced the error to < 1ms per update.

But that still meant you can get cumulative error. Since the number of updates can be a maximum of 4096 (going from off to fully on or vice versa), that means up to a maximum of 4096ms of cumulative error.

This change makes it possible to control how much cumulative error we build up by using units other than milliseconds, using the DispatchTimeInterval itself. We could use nanoseconds and reduce the error to about 4.1 microseconds, but if we use microseconds, the max cumulative error is 4.1 milliseconds. This is much better than 100th of a second, which is fine for the needs of this app.