Output Hardware PWM on GPIO 12/13 & 18/19
This enables the use of both sets of hardware PWM pins. 12 and 13 are a little bit easier to use when wiring (they both next to ground pins, unlike 19. Also unlike 19 which is used for SPI, 12 and 13 aren't used for any other purpose than GPIO.

Going forward, it might be a good idea to add something to the configuration that allows the user to pick high or low.