Align With Tweaked SingleBoard Interface
Fewer errors possible. Great.

Also get rid of the channel count logic in the config file. Not terribly useful. We’ll find out about bad channels when we attempt to create them with the PWM module, since that is what initializes the hardware for the output, not when we initialize the module. Fewer things for a user to track.