Reimplement Hardware Configuration on Codable
This is the first step towards moving away from Swift 3, by taking a dependency on Codable to handle the new hardware configuration JSON.

This doesn't make any changes that impact the running service for the moment. It lays the foundation for removing the SwiftyJSON dependency, and helping to split out the hardware and schedule configuration. This will be needed for MQTT support.