Massive Refactor of LED Output
This change makes a couple rather large changes:

—— Multiple Outputs:

It is now possible to specify multiple outputs in the hardware configuration JSON. Each hardware output is mapped to a single PWM output.

In the short term, there isn’t much guarding some user behavior like trying to use the Raspberry Pi’s PWM outputs twice. There definitely needs to be something to help block that down the road.

—— Named Channels:

Channels are now given names in the hardware configuration, and the schedule should refer to them directly, instead of things like “CHA01” or “PWM01” like before.

—— Refactored LED/PWM Control:

To support multiple LED modules at the same time, the whole PWM module has been refactored and cleaned up. It should make the code quite a bit simpler, cleaner, and smaller now that the amount of duplication for each module type is down from what it was before.

There’s also some new glue code work to make it a bit easier to have each module declare what configuration parameters it relies on, and to be able to throw errors if they happen to be missing.

Some of this needed to happen already, such as trimming the board types now that RPiLight uses SIngleBoard instead of SwiftyGPIO, because of SingleBoard’s better detection of which Raspberry Pi it is running on.