Add MCP4725 Output
This enables 0-10V output using things like the DAC boards available from Aptinex and ncd.io. This, at least on paper suggests I should be able to control a Kessil or similar light with two of these.

It also renames a couple of the files and the Impl class names to make them a bit more readable.