Fix Recursion Crash
There was an unfortunate recursion crash when trying to clamp the intensity to zero when it was being set at a value less than the minimum intensity. It would attempt to set the intensity to zero in a function called during intensity's didSet{} handler. Generally a bad idea. For now, let's go ahead and only clamp the values as they are sent out to the LEDModuleImpl.

This crept in during the refactor of the LED module.