Initial Network Module Work
Some initial bootstrapping work. Implements a couple key things:

1) Agnostic "Message" type. This has some nice benefits compared to PerfectMosquitto (that could be added via extensions), but can be converted back and forth. Performance of this isn't too important for this service.
2) Basic implementation of a Service.
3) Basic implementation of an Endpoint.

The idea here is that a Service routes incoming events to the correct Endpoints that handle those topics. Those endpoints route it then to specific closures. So it's a bit easier to handle multiple endpoints that might be owned by different parts of the daemon. One endpoint can be just for configuration, another for triggering preview behaviors, etc.

The implementation isn't complete, but it should be enough to prove out the core idea, and start looking towards building some read-only functionality so that an iOS app could begin to at least show information on aquariums it sees on the network.