Enable GitHub Actions for CI
* Basic CI does a very simple build for all pushes: Swift 5.0.3 and 5.1.1 on Ubuntu 18.04, and Xcode 11 on Mac. 
* Full CI adds Ubuntu 16.04 to the mix, as well as Xcode 10.3 on Mac. It also does builds on ARM Raspbian Stretch using QEMU. 

Unfortunately, because of some bugs with QEMU on 64-bit hosts when running 32-bit Debian Buster, that's not supported in the full CI at the moment. If those bugs were fixed, it would be possible (on paper) to do some Raspbian packages again, since we could build those using a workflow when a release is tagged.