Make Unit Tests Quieter
This provides more consistent quieting of the test output while running. Someone can easily change this locally if they want to see more logging during a particular test. But as the test count grows, this can get pretty ridiculous.

* Add Some Logging for Intermittent 5.0.3 Failure

Also move the QoS for the controller up to userInteractive, since that better fits the purpose of what the queue is doing.